link user and User's video & bug fix

+++ src/controllers/videoController.js
@@ -18,24 +18,37 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id } } = req.session;
     const video = await Video.findById(id);
+
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => { // 변경사항을 저장
+    const { user: { _id }, } = req.session;
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id }); // video object 대신 true or false를 받는 것
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
+    
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    
     await Video.findByIdAndUpdate(id, {
         title,
         description,
         hashtags: Video.formatHashtags(hashtags),
+        
     });
     return res.redirect(`/videos/${id}`);
 };
@@ -49,13 +62,17 @@ export const postUpload = async (req, res) => {
     const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
             owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+        return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", { pageTitle: "Upload Video", errorMessage: error._message });
     }
@@ -65,8 +82,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: { _id }, } = req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
-    // delete video
+    
     return res.redirect("/");
 };
 
@@ -81,4 +106,4 @@ export const search = async (req, res) => {
         });
     }
     return res.render("search", { pageTitle: "Search", videos });
-}
+};
